Moved all processing code into services/telemetry/processing
Telemetry class line count: 87. :smile:
